feat(upload): Рефакторинг завантаження файлів для динамічного отримання метаданих
Цей коміт значно переробляє механізм завантаження файлів, дозволяючи отримувати
ім'я файлу та загальний розмір безпосередньо у функції .
Це вирішує проблему необхідності  до того, як повні метадані файлу будуть доступні.

Основні зміни та покращення:
- **Єдина кінцева точка завантаження:** Замінено кінцеві точки  та  на єдину,
  потокову кінцеву точку , що спрощує взаємодію клієнт-сервер.
- **Динамічне отримання метаданих:** У  в :
    - Початковий запис файлу створюється в базі даних із тимчасовими метаданими
      для негайного отримання .
    - Фактичне  та  витягуються під час обробки
      багаточастинного потоку.
    - Після обробки всього файлу викликається нова функція
      для оновлення запису файлу з фінальним ,  та
      розрахованою .
- **Узгодженість типів даних у базі даних:** Оновлено тип  на  та  на
  у  та  для покращення обробки даних та узгодженості.
- **Покращення обробки частин:**
    - Виправлено регістр  ( -> ) при додаванні частин
      до черги, забезпечуючи правильну асоціацію.
    - Забезпечено, що остання часткова частина даних також коректно додається до черги для обробки.
    - Додано  після відправки до Telegram у  для звільнення пам'яті.
    - Введено  у  для можливого керування лімітами API Telegram.
- **Налаштування конфігурації:** Змінено  та розмір буфера
  для оптимізації продуктивності.

+++ db/structs.go
@@ -11,7 +11,7 @@ type File struct {
 	// TODO: додадти id, щоб можна було додавати до ключа
 	gorm.Model
 	FileName    string `json:"filename"`
-	Size        int    `json:"size"`
+	Size        int64  `json:"size"`
 	TotalChunks int
 	Status      string // uploading/completed/failed
 	OwnerAPIKey string `gorm:"index"`
@@ -23,9 +23,9 @@ type File struct {
 // It`s important to store the position to assamble the file
 type Chunk struct {
 	gorm.Model
-	FileID         int
+	FileID         uint
 	Position       int
-	Size           int
+	Size           int64
 	Status         string // pending/uploading/completed/failed
 	TelegramFileID string
 	Data           []byte
